# incubation/node-manager/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/node-manager/internal/cloud-region/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CloudRegion"
                ],
                "summary": "创建云区域",
                "parameters": [
                    {
                        "description": "entity",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CloudRegionCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    }
                }
            }
        },
        "/node-manager/internal/cloud-region/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CloudRegion"
                ],
                "summary": "获取所有云区域",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.CloudRegionItemResponse"
                            }
                        }
                    }
                }
            }
        },
        "/node-manager/internal/cloud-region/update": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CloudRegion"
                ],
                "summary": "更新云区域",
                "parameters": [
                    {
                        "description": "entity",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CloudRegionUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    }
                }
            }
        },
        "/node-manager/internal/variable/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variable"
                ],
                "summary": "创建变量",
                "parameters": [
                    {
                        "description": "entity",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.VariableCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    }
                }
            }
        },
        "/node-manager/internal/variable/delete/multiple": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variable"
                ],
                "summary": "删除多个变量",
                "parameters": [
                    {
                        "description": "Variable IDs",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    }
                }
            }
        },
        "/node-manager/internal/variable/delete/single/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variable"
                ],
                "summary": "删除单个变量",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Variable ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    }
                }
            }
        },
        "/node-manager/internal/variable/list/{cloudRegionId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variable"
                ],
                "summary": "获取该云区域下的所有变量",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cloud Region ID",
                        "name": "cloudRegionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.VariableItemResponse"
                            }
                        }
                    }
                }
            }
        },
        "/node-manager/internal/variable/search/{cloudRegionId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variable"
                ],
                "summary": "搜索该云区域下的变量",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cloud Region ID",
                        "name": "cloudRegionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search Value",
                        "name": "value",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.VariableItemResponse"
                            }
                        }
                    }
                }
            }
        },
        "/node-manager/internal/variable/update": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variable"
                ],
                "summary": "修改变量",
                "parameters": [
                    {
                        "description": "entity",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.VariableUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.CloudRegionCreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "introduction": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.CloudRegionItemResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "introduction": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.CloudRegionUpdateRequest": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "introduction": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.VariableCreateRequest": {
            "type": "object",
            "required": [
                "name",
                "value"
            ],
            "properties": {
                "cloud_region_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "entity.VariableItemResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "entity.VariableUpdateRequest": {
            "type": "object",
            "required": [
                "id",
                "name",
                "value"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}